+++ src/yyjson.c
@@ -3688,7 +3688,7 @@ static_inline bool read_inf_or_nan(bool sign, u8 **ptr, u8 **pre,
 /** Read a JSON number as raw string. */
 static_noinline bool read_number_raw(u8 **ptr,
                                      u8 **pre,
-                                     bool ext,
+                                     yyjson_read_flag flg,
                                      yyjson_val *val,
                                      const char **msg) {
     
@@ -3716,7 +3716,7 @@ static_noinline bool read_number_raw(u8 **ptr,
     
     /* read first digit, check leading zero */
     if (unlikely(!digi_is_digit(*cur))) {
-        if (unlikely(ext)) {
+        if (flg & YYJSON_READ_ALLOW_INF_AND_NAN) {
             if (read_inf_or_nan(*hdr == '-', &cur, pre, val)) return_raw();
         }
         return_err(cur, "no digit after minus sign");
@@ -4200,7 +4200,7 @@ static const f64 f64_pow10_table[] = {
  */
 static_inline bool read_number(u8 **ptr,
                                u8 **pre,
-                               bool ext,
+                               yyjson_read_flag flg,
                                yyjson_val *val,
                                const char **msg) {
     
@@ -4228,15 +4228,23 @@ static_inline bool read_number(u8 **ptr,
     *end = cur; return true; \
 } while (false)
     
-#define return_f64_raw(_v) do { \
+#define return_f64_bin(_v) do { \
     val->tag = YYJSON_TYPE_NUM | YYJSON_SUBTYPE_REAL; \
     val->uni.u64 = ((u64)sign << 63) | (u64)(_v); \
     *end = cur; return true; \
 } while (false)
     
 #define return_inf() do { \
-    if (unlikely(ext)) return_f64_raw(F64_RAW_INF); \
+    if (flg & YYJSON_READ_BIGNUM_AS_RAW) return_raw(); \
+    if (flg & YYJSON_READ_ALLOW_INF_AND_NAN) return_f64_bin(F64_RAW_INF); \
     else return_err(hdr, "number is infinity when parsed as double"); \
+} while (false)
+    
+#define return_raw() do { \
+    if (*pre) **pre = '\0'; /* add null-terminator for previous raw string */ \
+    val->tag = ((u64)(cur - hdr) << YYJSON_TAG_BIT) | YYJSON_TYPE_RAW; \
+    val->uni.str = (const char *)hdr; \
+    *pre = cur; *end = cur; return true; \
 } while (false)
     
     u8 *sig_cut = NULL; /* significant part cutting position for long number */
@@ -4257,9 +4265,9 @@ static_inline bool read_number(u8 **ptr,
     u8 **end = ptr;
     bool sign;
     
-    /* read number as raw string if has flag */
-    if (unlikely(pre)) {
-        return read_number_raw(ptr, pre, ext, val, msg);
+    /* read number as raw string if has `YYJSON_READ_NUMBER_AS_RAW` flag */
+    if (unlikely(pre && !(flg & YYJSON_READ_BIGNUM_AS_RAW))) {
+        return read_number_raw(ptr, pre, flg, val, msg);
     }
     
     sign = (*hdr == '-');
@@ -4268,7 +4276,7 @@ static_inline bool read_number(u8 **ptr,
     /* begin with a leading zero or non-digit */
     if (unlikely(!digi_is_nonzero(*cur))) { /* 0 or non-digit char */
         if (unlikely(*cur != '0')) { /* non-digit char */
-            if (unlikely(ext)) {
+            if (flg & YYJSON_READ_ALLOW_INF_AND_NAN) {
                 if (read_inf_or_nan(sign, &cur, pre, val)) {
                     *end = cur;
                     return true;
@@ -4301,7 +4309,7 @@ static_inline bool read_number(u8 **ptr,
             }
             while (digi_is_digit(*++cur));
         }
-        return_f64_raw(0);
+        return_f64_bin(0);
     }
     
     /* begin with non-zero digit */
@@ -4327,6 +4335,7 @@ static_inline bool read_number(u8 **ptr,
     if (!digi_is_digit_or_fp(*cur)) {
         /* this number is an integer consisting of 19 digits */
         if (sign && (sig > ((u64)1 << 63))) { /* overflow */
+            if (flg & YYJSON_READ_BIGNUM_AS_RAW) return_raw();
             return_f64(normalized_u64_to_f64(sig));
         }
         return_i64(sig);
@@ -4379,7 +4388,10 @@ digi_intg_more:
                 sig = num + sig * 10;
                 cur++;
                 /* convert to double if overflow */
-                if (sign) return_f64(normalized_u64_to_f64(sig));
+                if (sign) {
+                    if (flg & YYJSON_READ_BIGNUM_AS_RAW) return_raw();
+                    return_f64(normalized_u64_to_f64(sig));
+                }
                 return_i64(sig);
             }
         }
@@ -4404,6 +4416,9 @@ digi_frac_more:
     sig += (*cur >= '5'); /* round */
     while (digi_is_digit(*++cur));
     if (!dot_pos) {
+        if (!digi_is_fp(*cur) && (flg & YYJSON_READ_BIGNUM_AS_RAW)) {
+            return_raw(); /* it's a large integer */
+        }
         dot_pos = cur;
         if (*cur == '.') {
             if (!digi_is_digit(*++cur)) {
@@ -4437,7 +4452,7 @@ digi_frac_end:
     exp_sig = -(i64)((u64)(cur - dot_pos) - 1);
     if (likely(!digi_is_exp(*cur))) {
         if (unlikely(exp_sig < F64_MIN_DEC_EXP - 19)) {
-            return_f64_raw(0); /* underflow */
+            return_f64_bin(0); /* underflow */
         }
         exp = (i32)exp_sig;
         goto digi_finish;
@@ -4462,7 +4477,7 @@ digi_exp_more:
     }
     if (unlikely(cur - tmp >= U64_SAFE_DIG)) {
         if (exp_sign) {
-            return_f64_raw(0); /* underflow */
+            return_f64_bin(0); /* underflow */
         } else {
             return_inf(); /* overflow */
         }
@@ -4473,7 +4488,7 @@ digi_exp_more:
     /* validate exponent value */
 digi_exp_finish:
     if (unlikely(exp_sig < F64_MIN_DEC_EXP - 19)) {
-        return_f64_raw(0); /* underflow */
+        return_f64_bin(0); /* underflow */
     }
     if (unlikely(exp_sig > F64_MAX_DEC_EXP)) {
         return_inf(); /* overflow */
@@ -4644,7 +4659,7 @@ digi_finish:
             exp2 += F64_BITS - F64_SIG_FULL_BITS + F64_SIG_BITS;
             exp2 += F64_EXP_BIAS;
             raw = ((u64)exp2 << F64_SIG_BITS) | (hi & F64_SIG_MASK);
-            return_f64_raw(raw);
+            return_f64_bin(raw);
         }
     }
     
@@ -4734,7 +4749,7 @@ digi_finish:
         if (unlikely(raw == F64_RAW_INF)) return_inf();
         if (likely(precision_bits <= half_way - fp_err ||
                    precision_bits >= half_way + fp_err)) {
-            return_f64_raw(raw); /* number is accurate */
+            return_f64_bin(raw); /* number is accurate */
         }
         /* now the number is the correct value, or the next lower value */
         
@@ -4774,16 +4789,16 @@ digi_finish:
         }
         
         if (unlikely(raw == F64_RAW_INF)) return_inf();
-        return_f64_raw(raw);
+        return_f64_bin(raw);
     }
     
-#undef has_flag
 #undef return_err
 #undef return_inf
 #undef return_0
 #undef return_i64
 #undef return_f64
-#undef return_f64_raw
+#undef return_f64_bin
+#undef return_raw
 }
 
 
@@ -4797,7 +4812,7 @@ digi_finish:
  */
 static_noinline bool read_number(u8 **ptr,
                                  u8 **pre,
-                                 bool ext,
+                                 yyjson_read_flag flg,
                                  yyjson_val *val,
                                  const char **msg) {
     
@@ -4825,10 +4840,23 @@ static_noinline bool read_number(u8 **ptr,
     *end = cur; return true; \
 } while (false)
     
-#define return_f64_raw(_v) do { \
+#define return_f64_bin(_v) do { \
     val->tag = YYJSON_TYPE_NUM | YYJSON_SUBTYPE_REAL; \
     val->uni.u64 = ((u64)sign << 63) | (u64)(_v); \
     *end = cur; return true; \
+} while (false)
+    
+#define return_inf() do { \
+    if (flg & YYJSON_READ_BIGNUM_AS_RAW) return_raw(); \
+    if (flg & YYJSON_READ_ALLOW_INF_AND_NAN) return_f64_bin(F64_RAW_INF); \
+    else return_err(hdr, "number is infinity when parsed as double"); \
+} while (false)
+    
+#define return_raw() do { \
+    if (*pre) **pre = '\0'; /* add null-terminator for previous raw string */ \
+    val->tag = ((u64)(cur - hdr) << YYJSON_TAG_BIT) | YYJSON_TYPE_RAW; \
+    val->uni.str = (const char *)hdr; \
+    *pre = cur; *end = cur; return true; \
 } while (false)
     
     u64 sig, num;
@@ -4839,9 +4867,9 @@ static_noinline bool read_number(u8 **ptr,
     u8 *f64_end = NULL;
     bool sign;
     
-    /* read number as raw string if has flag */
-    if (unlikely(pre)) {
-        return read_number_raw(ptr, pre, ext, val, msg);
+    /* read number as raw string if has `YYJSON_READ_NUMBER_AS_RAW` flag */
+    if (unlikely(pre && !(flg & YYJSON_READ_BIGNUM_AS_RAW))) {
+        return read_number_raw(ptr, pre, flg, val, msg);
     }
     
     sign = (*hdr == '-');
@@ -4850,7 +4878,7 @@ static_noinline bool read_number(u8 **ptr,
     
     /* read first digit, check leading zero */
     if (unlikely(!digi_is_digit(*cur))) {
-        if (unlikely(ext)) {
+        if (flg & YYJSON_READ_ALLOW_INF_AND_NAN) {
             if (read_inf_or_nan(sign, &cur, pre, val)) {
                 *end = cur;
                 return true;
@@ -4883,7 +4911,10 @@ static_noinline bool read_number(u8 **ptr,
             (sig == (U64_MAX / 10) && num <= (U64_MAX % 10))) {
             sig = num + sig * 10;
             cur++;
-            if (sign) return_f64(normalized_u64_to_f64(sig));
+            if (sign) {
+                if (flg & YYJSON_READ_BIGNUM_AS_RAW) return_raw();
+                return_f64(normalized_u64_to_f64(sig));
+            }
             return_i64(sig);
         }
     }
@@ -4893,6 +4924,7 @@ intg_end:
     if (!digi_is_digit_or_fp(*cur)) {
         /* this number is an integer consisting of 1 to 19 digits */
         if (sign && (sig > ((u64)1 << 63))) {
+            if (flg & YYJSON_READ_BIGNUM_AS_RAW) return_raw();
             return_f64(normalized_u64_to_f64(sig));
         }
         return_i64(sig);
@@ -4901,6 +4933,9 @@ intg_end:
 read_double:
     /* this number should be read as double */
     while (digi_is_digit(*cur)) cur++;
+    if (!digi_is_fp(*cur) && (flg & YYJSON_READ_BIGNUM_AS_RAW)) {
+        return_raw(); /* it's a large integer */
+    }
     if (*cur == '.') {
         /* skip fraction part */
         dot = cur;
@@ -4935,29 +4970,32 @@ read_double:
      */
     val->uni.f64 = strtod((const char *)hdr, (char **)&f64_end);
     if (unlikely(f64_end != cur)) {
+        /* replace '.' with ',' for locale */
         bool cut = (*cur == ',');
-        if (dot) *dot = ',';
         if (cut) *cur = ' ';
+        if (dot) *dot = ',';
         val->uni.f64 = strtod((const char *)hdr, (char **)&f64_end);
+        /* restore ',' to '.' */
         if (cut) *cur = ',';
+        if (dot) *dot = '.';
         if (unlikely(f64_end != cur)) {
             return_err(hdr, "strtod() failed to parse the number");
         }
     }
     if (unlikely(val->uni.f64 >= HUGE_VAL || val->uni.f64 <= -HUGE_VAL)) {
-        if (!ext) {
-            return_err(hdr, "number is infinity when parsed as double");
-        }
+        return_inf();
     }
     val->tag = YYJSON_TYPE_NUM | YYJSON_SUBTYPE_REAL;
     *end = cur;
     return true;
     
-#undef has_flag
 #undef return_err
 #undef return_0
 #undef return_i64
 #undef return_f64
+#undef return_f64_bin
+#undef return_inf
+#undef return_raw
 }
 
 #endif /* FP_READER */
@@ -5464,7 +5502,6 @@ static_noinline yyjson_doc *read_root_single(u8 *hdr,
     const char *msg; /* error message */
     
     bool raw; /* read number as raw */
-    bool ext; /* allow inf and nan */
     bool inv; /* allow invalid unicode */
     u8 *raw_end; /* raw end for null-terminator */
     u8 **pre; /* previous raw end pointer */
@@ -5476,14 +5513,13 @@ static_noinline yyjson_doc *read_root_single(u8 *hdr,
     val_hdr = (yyjson_val *)alc.malloc(alc.ctx, alc_num * sizeof(yyjson_val));
     if (unlikely(!val_hdr)) goto fail_alloc;
     val = val_hdr + hdr_len;
-    raw = (flg & YYJSON_READ_NUMBER_AS_RAW) != 0;
-    ext = (flg & YYJSON_READ_ALLOW_INF_AND_NAN) != 0;
+    raw = (flg & (YYJSON_READ_NUMBER_AS_RAW | YYJSON_READ_BIGNUM_AS_RAW)) != 0;
     inv = (flg & YYJSON_READ_ALLOW_INVALID_UNICODE) != 0;
     raw_end = NULL;
     pre = raw ? &raw_end : NULL;
     
     if (char_is_number(*cur)) {
-        if (likely(read_number(&cur, pre, ext, val, &msg))) goto doc_end;
+        if (likely(read_number(&cur, pre, flg, val, &msg))) goto doc_end;
         goto fail_number;
     }
     if (*cur == '"') {
@@ -5500,12 +5536,12 @@ static_noinline yyjson_doc *read_root_single(u8 *hdr,
     }
     if (*cur == 'n') {
         if (likely(read_null(&cur, val))) goto doc_end;
-        if (unlikely(ext)) {
+        if (unlikely(flg & YYJSON_READ_ALLOW_INF_AND_NAN)) {
             if (read_nan(false, &cur, pre, val)) goto doc_end;
         }
         goto fail_literal;
     }
-    if (unlikely(ext)) {
+    if (unlikely(flg & YYJSON_READ_ALLOW_INF_AND_NAN)) {
         if (read_inf_or_nan(false, &cur, pre, val)) goto doc_end;
     }
     goto fail_character;
@@ -5607,7 +5643,6 @@ static_inline yyjson_doc *read_root_minify(u8 *hdr,
     const char *msg; /* error message */
     
     bool raw; /* read number as raw */
-    bool ext; /* allow inf and nan */
     bool inv; /* allow invalid unicode */
     u8 *raw_end; /* raw end for null-terminator */
     u8 **pre; /* previous raw end pointer */
@@ -5625,8 +5660,7 @@ static_inline yyjson_doc *read_root_minify(u8 *hdr,
     val = val_hdr + hdr_len;
     ctn = val;
     ctn_len = 0;
-    raw = (flg & YYJSON_READ_NUMBER_AS_RAW) != 0;
-    ext = (flg & YYJSON_READ_ALLOW_INF_AND_NAN) != 0;
+    raw = (flg & (YYJSON_READ_NUMBER_AS_RAW | YYJSON_READ_BIGNUM_AS_RAW)) != 0;
     inv = (flg & YYJSON_READ_ALLOW_INVALID_UNICODE) != 0;
     raw_end = NULL;
     pre = raw ? &raw_end : NULL;
@@ -5667,7 +5701,7 @@ arr_val_begin:
     if (char_is_number(*cur)) {
         val_incr();
         ctn_len++;
-        if (likely(read_number(&cur, pre, ext, val, &msg))) goto arr_val_end;
+        if (likely(read_number(&cur, pre, flg, val, &msg))) goto arr_val_end;
         goto fail_number;
     }
     if (*cur == '"') {
@@ -5692,7 +5726,7 @@ arr_val_begin:
         val_incr();
         ctn_len++;
         if (likely(read_null(&cur, val))) goto arr_val_end;
-        if (unlikely(ext)) {
+        if (unlikely(flg & YYJSON_READ_ALLOW_INF_AND_NAN)) {
             if (read_nan(false, &cur, pre, val)) goto arr_val_end;
         }
         goto fail_literal;
@@ -5708,7 +5742,8 @@ arr_val_begin:
         while (char_is_space(*++cur));
         goto arr_val_begin;
     }
-    if (unlikely(ext) && (*cur == 'i' || *cur == 'I' || *cur == 'N')) {
+    if (unlikely(flg & YYJSON_READ_ALLOW_INF_AND_NAN) &&
+        (*cur == 'i' || *cur == 'I' || *cur == 'N')) {
         val_incr();
         ctn_len++;
         if (read_inf_or_nan(false, &cur, pre, val)) goto arr_val_end;
@@ -5817,7 +5852,7 @@ obj_val_begin:
     if (char_is_number(*cur)) {
         val++;
         ctn_len++;
-        if (likely(read_number(&cur, pre, ext, val, &msg))) goto obj_val_end;
+        if (likely(read_number(&cur, pre, flg, val, &msg))) goto obj_val_end;
         goto fail_number;
     }
     if (*cur == '{') {
@@ -5844,7 +5879,7 @@ obj_val_begin:
         val++;
         ctn_len++;
         if (likely(read_null(&cur, val))) goto obj_val_end;
-        if (unlikely(ext)) {
+        if (unlikely(flg & YYJSON_READ_ALLOW_INF_AND_NAN)) {
             if (read_nan(false, &cur, pre, val)) goto obj_val_end;
         }
         goto fail_literal;
@@ -5853,7 +5888,8 @@ obj_val_begin:
         while (char_is_space(*++cur));
         goto obj_val_begin;
     }
-    if (unlikely(ext) && (*cur == 'i' || *cur == 'I' || *cur == 'N')) {
+    if (unlikely(flg & YYJSON_READ_ALLOW_INF_AND_NAN) &&
+        (*cur == 'i' || *cur == 'I' || *cur == 'N')) {
         val++;
         ctn_len++;
         if (read_inf_or_nan(false, &cur, pre, val)) goto obj_val_end;
@@ -5997,7 +6033,6 @@ static_inline yyjson_doc *read_root_pretty(u8 *hdr,
     const char *msg; /* error message */
     
     bool raw; /* read number as raw */
-    bool ext; /* allow inf and nan */
     bool inv; /* allow invalid unicode */
     u8 *raw_end; /* raw end for null-terminator */
     u8 **pre; /* previous raw end pointer */
@@ -6015,8 +6050,7 @@ static_inline yyjson_doc *read_root_pretty(u8 *hdr,
     val = val_hdr + hdr_len;
     ctn = val;
     ctn_len = 0;
-    raw = (flg & YYJSON_READ_NUMBER_AS_RAW) != 0;
-    ext = (flg & YYJSON_READ_ALLOW_INF_AND_NAN) != 0;
+    raw = (flg & (YYJSON_READ_NUMBER_AS_RAW | YYJSON_READ_BIGNUM_AS_RAW)) != 0;
     inv = (flg & YYJSON_READ_ALLOW_INVALID_UNICODE) != 0;
     raw_end = NULL;
     pre = raw ? &raw_end : NULL;
@@ -6072,7 +6106,7 @@ arr_val_begin:
     if (char_is_number(*cur)) {
         val_incr();
         ctn_len++;
-        if (likely(read_number(&cur, pre, ext, val, &msg))) goto arr_val_end;
+        if (likely(read_number(&cur, pre, flg, val, &msg))) goto arr_val_end;
         goto fail_number;
     }
     if (*cur == '"') {
@@ -6097,7 +6131,7 @@ arr_val_begin:
         val_incr();
         ctn_len++;
         if (likely(read_null(&cur, val))) goto arr_val_end;
-        if (unlikely(ext)) {
+        if (unlikely(flg & YYJSON_READ_ALLOW_INF_AND_NAN)) {
             if (read_nan(false, &cur, pre, val)) goto arr_val_end;
         }
         goto fail_literal;
@@ -6113,7 +6147,8 @@ arr_val_begin:
         while (char_is_space(*++cur));
         goto arr_val_begin;
     }
-    if (unlikely(ext) && (*cur == 'i' || *cur == 'I' || *cur == 'N')) {
+    if (unlikely(flg & YYJSON_READ_ALLOW_INF_AND_NAN) &&
+        (*cur == 'i' || *cur == 'I' || *cur == 'N')) {
         val_incr();
         ctn_len++;
         if (read_inf_or_nan(false, &cur, pre, val)) goto arr_val_end;
@@ -6243,7 +6278,7 @@ obj_val_begin:
     if (char_is_number(*cur)) {
         val++;
         ctn_len++;
-        if (likely(read_number(&cur, pre, ext, val, &msg))) goto obj_val_end;
+        if (likely(read_number(&cur, pre, flg, val, &msg))) goto obj_val_end;
         goto fail_number;
     }
     if (*cur == '{') {
@@ -6270,7 +6305,7 @@ obj_val_begin:
         val++;
         ctn_len++;
         if (likely(read_null(&cur, val))) goto obj_val_end;
-        if (unlikely(ext)) {
+        if (unlikely(flg & YYJSON_READ_ALLOW_INF_AND_NAN)) {
             if (read_nan(false, &cur, pre, val)) goto obj_val_end;
         }
         goto fail_literal;
@@ -6279,7 +6314,8 @@ obj_val_begin:
         while (char_is_space(*++cur));
         goto obj_val_begin;
     }
-    if (unlikely(ext) && (*cur == 'i' || *cur == 'I' || *cur == 'N')) {
+    if (unlikely(flg & YYJSON_READ_ALLOW_INF_AND_NAN) &&
+        (*cur == 'i' || *cur == 'I' || *cur == 'N')) {
         val++;
         ctn_len++;
         if (read_inf_or_nan(false, &cur, pre, val)) goto obj_val_end;
@@ -6632,7 +6668,6 @@ const char *yyjson_read_number(const char *dat,
     
     u8 *hdr = constcast(u8 *)dat, *cur = hdr;
     bool raw; /* read number as raw */
-    bool ext; /* allow inf and nan */
     u8 *raw_end; /* raw end for null-terminator */
     u8 **pre; /* previous raw end pointer */
     const char *msg;
@@ -6666,28 +6701,27 @@ const char *yyjson_read_number(const char *dat,
         }
         memcpy(hdr, dat, dat_len + 1);
     }
+    hdr[dat_len] = 0;
 #endif
     
 #if YYJSON_DISABLE_NON_STANDARD
-    ext = false;
-#else
-    ext = (flg & YYJSON_READ_ALLOW_INF_AND_NAN) != 0;
+    flg &= ~YYJSON_READ_ALLOW_INF_AND_NAN;
 #endif
     
-    raw = (flg & YYJSON_READ_NUMBER_AS_RAW) != 0;
+    raw = (flg & (YYJSON_READ_NUMBER_AS_RAW | YYJSON_READ_BIGNUM_AS_RAW)) != 0;
     raw_end = NULL;
     pre = raw ? &raw_end : NULL;
     
 #if !YYJSON_HAS_IEEE_754 || YYJSON_DISABLE_FAST_FP_CONV
-    if (!read_number(&cur, pre, ext, val, &msg)) {
+    if (!read_number(&cur, pre, flg, val, &msg)) {
         if (dat_len >= sizeof(buf)) alc->free(alc->ctx, hdr);
         return_err(cur, INVALID_NUMBER, msg);
     }
     if (dat_len >= sizeof(buf)) alc->free(alc->ctx, hdr);
-    if (raw) val->uni.str = dat;
+    if (yyjson_is_raw(val)) val->uni.str = dat;
     return dat + (cur - hdr);
 #else
-    if (!read_number(&cur, pre, ext, val, &msg)) {
+    if (!read_number(&cur, pre, flg, val, &msg)) {
         return_err(cur, INVALID_NUMBER, msg);
     }
     return (const char *)cur;